Initial module to interact with the android bcb.
Android uses a dedicated partition or mtd to handle communication
between the bootloader and the recovery system, which handles system
and firmware updates as well as factory resets. This is known as the
bootloader control block, or "bcb".

The current form of the module allows the user to specify which
partition should be used as the bcb, as well as reading the 'command'
field of the message via the android_bcb_command env var. During a
recovery operation, the recovery system populates the command field and
only clears it once the operation is done, with the intended semantics
being that the system reboots into recovery mode if the operation is
interrupted. Thus, a grub configuration can check to see if
android_bcb_command non-empty, and if so boot into recovery.

Future versions may give access to other fields, or allow write access.